requirements.txt
arrow==1.3.0
asgiref==3.9.1
beautifulsoup4==4.13.4
binaryornot==0.4.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
cssbeautifier==1.15.4
Django==5.2.4
django-browser-reload==1.18.0
django-debug-toolbar==6.0.0
django-extensions==4.1
django-htmx==1.23.2
django-render-block==0.11
django-tailwind==4.2.0
django-template-partials==25.1
django-widget-tweaks==1.5.0
djlint==1.36.4
EditorConfig==0.17.1
Faker==37.5.3
honcho==2.0.0
idna==3.10
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pathspec==0.12.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
